Perform submission reading and replying in same loops
Before we had the program read the recent submissions into memory, storing all the relevant object info into a dict, and then replying. I thought this would be cleaner, but it makes accessing the Comment object we need to reply to awkward, and I think we can eventually get it to be relatively close to the same cleanliness. Additionally, this lets us perform all the related functions one after another: read the comment, reply to it, and log the reply.