Make the continue conditional clearer + fix it
Use (p OR q) instead of ~(p AND q). Shouldn't have stayed in there this long, must have been an artifact from an old commit. Also, the check that a post should be newer than last Sunday was incorrect, and I didn't notice because I've been testing the bot in a subreddit that's a few days old. P.S., 'newer than last Sunday' is probably too strict, since I assume that some posts from late on Sunday don't make it in until the subsequent week. TODO follow up on this